+++ R/show_earning_panel.R
@@ -1,11 +1,11 @@
 #' Show earning panel
 #' @param title character string with title of panel
 #' @param outputId outputId of tabel output
-#' 
-show_earning_panel <- function(title, outputId) {
+#' @param ... additional arguments applied to shiny::tabPanel
+show_earning_panel <- function(title, outputId, ...) {
   
   tabPanel(title = title,
            br(),
-           DT::dataTableOutput(outputId = outputId))
+           DT::dataTableOutput(outputId = outputId), ...)
   
 }